# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ypk btpit",
        "description": "Ypk btpit",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Пользователь"
                ],
                "summary": "Получить информацию о текущем пользователе",
                "operationId": "a78fa60870f290ed5ec1d89eb5df741c",
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "phoneNumber": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "role": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/me/all": {
            "get": {
                "tags": [
                    "Пользователь"
                ],
                "summary": "Получить полную информацию о текущем пользователе",
                "operationId": "bf31440a28c182b01e1782ac6bc23243",
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "phoneNumber": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "role": {
                                            "type": "string"
                                        },
                                        "userInfo": {
                                            "type": "string"
                                        },
                                        "avatarUrl": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}